rev-list: drop unused void pointer from finish_commit()
Our finish_commit() function used to be passed directly to the revision
machinery as a callback. But after f34c8cd0e8 (rev-list: fix
--verify-objects --quiet becoming --objects, 2012-02-28), it is used
only as a helper in show_commit().

It doesn't use its void "data" parameter, and we no longer have to
conform to the callback interface. Let's drop it.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>